requirement.txt
PySastrawi